move fe to web folder, add backtrack, single point delete and class based plot
