--- src/utils/PlotInitializer.js
@@ -1,18 +1,0 @@
-class PlotInitializer {
-  constructor(data, shownInstructionData, shownOutputData, isTextSearching) {
-    this.data = data;
-    this.shownInstructionData = shownInstructionData;
-    this.shownOutputData = shownOutputData;
-    this.isTextSearching = isTextSearching;
-  }
-
-  async initPlot() {
-    // Your existing code for the initPlot function goes here
-    // Remember to replace references to variables with this.variableName
-    // For example, replace data with this.data, shownInstructionData with this.shownInstructionData, etc.
-  }
-}
-
-// Usage:
-// const plotInitializer = new PlotInitializer(data, shownInstructionData, shownOutputData, isTextSearching);
-// plotInitializer.initPlot();
