+++ web/tailwind.config.js
